[tapilite] add TAPI emulation library
TAPI is a project with close ties to LLVM, which means one needs it to
build a linker.
On systems where no LLVM is used (like older Darwin), or systems where
GCC is in use instead, this means TAPI cannot be built.
On recent Darwin systems, a linker that does not do TAPI, cannot do
much, since all system libraries are .tbd (TAPI stubs).

Hence, to e.g. bootstrap LLVM, we need to be able to build a linker that
understands TAPI, without having TAPI (the host may not supply it
either).

This introduces TAPIlite, an implementation of those classes ld64 uses,
with help of libyaml.  This is very limited at this point, and whether
or not it works is very brittle.  Further development here is likely
necessary.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>